separated Multer function into routes and controllers
